Add functionality to remove context elements by name
Added a new function `ellama-context--element-remove-by-name` to remove all
context elements that match by name. Also added an interactive function
`ellama-context-element-remove-by-name` to prompt the user for the name of the
element to remove and then call the internal function. Updated the transient
prefix menu to include a new "Delete element" option under the "Manage" section.