Ensure buffer exists before updating context
Ensured that the buffer `ellama--context-buffer` is created if it does not exist
before attempting to update and show context. This prevents errors related to
non-existent buffers. Also, added calls to `ellama-update-context-show` in both
header line and mode line minor modes to ensure context is shown when these
modes are toggled.