Add transient menu for blueprint mode commands
Added a new transient prefix command `ellama-transient-blueprint-mode-menu` to
provide a menu of options for interacting with the current blueprint buffer.
This includes sending the buffer to chat, setting a system message, creating a
new blueprint, filling variables, and quitting or killing the buffer. Updated
keybindings in `ellama-blueprint-mode-map` to use this new transient command.
Also added a helper function `ellama-chat-with-system-from-buffer` to start a
chat session with a system message created from the current buffer content.