Add ephemeral session option to transient commands
Added the `--ephemeral` option to several transient commands in
`ellama-transient.el`. This allows users to specify whether they want an
ephemeral session when triggering code review, asking questions, or chatting
with Ellama. Also updated `ellama.el` to ensure that auto-saving only occurs if
both `ellama-session-auto-save` is enabled and the buffer has a file name
associated with it.

Fix #283