Remove `ellama-ollama-binary` configuration
Removed the unused `ellama-ollama-binary` custom variable and related code from
`README.org`, `ellama.el`, and `ellama.info`. This cleanup simplifies the
configuration and removes redundant logic. Given that Ollama's API nature makes
it possible, we should consider the feasibility of directly interfacing with its
API instead of relying on a local Ollama installation. This approach would be
particularly beneficial for individuals who host Ollama on a separate server, as
it would eliminate the need for Ollama to be installed on their client machine.

Fix #228