ellama-blueprints: be less aggressive matching template vars
As various languages use {} as function markers its hard to include
example code snippets in blueprints due to the variable expansion. It
seems reasonable to restrict the substitutions to {selfcontained}
variables that don't span multiple lines.

While at it I created a variable for the regexp so future changes
don't need to find all the references.